refactor(cn0): split the CN0 estimators into one file each
`src/cn0_estimation.jl` held four estimators plus the shared abstraction in
737 lines, and a fifth is about to arrive. Mirror the `src/correlators/`
convention instead — a folder whose shared file is the singular of the folder
name, then one file per concrete type:

    src/cn0_estimators/cn0_estimator.jl   AbstractCN0Estimator, CN0UpdateContext,
                                          the 3-arg `update` extension point,
                                          default_cn0_estimator
    src/cn0_estimators/moments.jl         MomentsCN0Estimator
    src/cn0_estimators/no_cn0.jl          NoCN0Estimator
    src/cn0_estimators/nwpr.jl            NWPRCN0Estimator + its window helpers

Each move range includes the type's own docstring — `NWPRCN0Estimator`'s alone
is 160 lines, so splitting it off would have left the type's documentation
stranded in the shared file.

Pure motion: no behaviour, name, signature or docstring change. The shared file
still follows `bit_buffer.jl`, since `CN0UpdateContext` has a `BitBuffer` field;
`default_cn0_estimator` only references the concrete estimators inside a
function body, so it resolves at call time and stays in the shared file.

`test/cn0_estimation.jl` follows the same split into `test/cn0_estimators/`,
each file its own module with an explicit import list.
