+++ test/cn0_estimators/moments.jl
@@ -1,0 +1,90 @@
+module MomentsCN0EstimatorTest
+
+using Test: @test, @testset, @inferred
+using Random: Random
+using Unitful: kHz, MHz, Hz, ms, dBHz
+using StaticArrays: SVector
+using GNSSSignals: GPSL1CA, get_code
+import Tracking
+using Tracking:
+    MomentsCN0Estimator,
+    get_prompt_buffer,
+    get_current_index,
+    update,
+    EarlyPromptLateCorrelator,
+    get_prompt,
+    get_correlator_sample_shifts,
+    estimate_cn0
+
+@testset "Moments CN0 estimator" begin
+    cn0_estimator = MomentsCN0Estimator(20)
+    @test @inferred(get_prompt_buffer(cn0_estimator)) == zero(SVector{20,ComplexF64})
+    @test @inferred(get_current_index(cn0_estimator)) == 0
+    @test @inferred(Base.length(cn0_estimator)) == 0
+
+    next_cn0_estimator = @inferred update(cn0_estimator, 1 + 2im)
+    @test @inferred(get_prompt_buffer(next_cn0_estimator))[1] == 1 + 2im
+    @test @inferred(get_current_index(next_cn0_estimator)) == 1
+    @test @inferred(Base.length(next_cn0_estimator)) == 1
+
+    cn0_estimator = MomentsCN0Estimator(ones(SVector{20,ComplexF64}), 20, 20)
+    next_cn0_estimator = @inferred update(cn0_estimator, 1 + 2im)
+    @test @inferred(get_prompt_buffer(next_cn0_estimator))[1] == 1 + 2im
+    @test @inferred(get_current_index(next_cn0_estimator)) == 1
+    @test @inferred(Base.length(next_cn0_estimator)) == 20
+
+    cn0_estimator = MomentsCN0Estimator(ones(SVector{20,ComplexF64}), 19, 20)
+    @test @inferred(get_current_index(cn0_estimator)) == 19
+    @test @inferred(Base.length(cn0_estimator)) == 20
+end
+
+@testset "CN0 estimation" begin
+    Random.seed!(1234)
+    carrier_doppler = 0Hz
+    start_code_phase = 0
+    code_frequency = 1023kHz
+    sampling_frequency = 4MHz
+    prn = 1
+    range = 0:3999
+    start_carrier_phase = π / 2
+    cn0_estimator = MomentsCN0Estimator(100)
+    start_sample = 1
+    num_samples = 4000
+    gpsl1 = GPSL1CA()
+
+    for i = 1:100
+        signal =
+            get_code.(
+                gpsl1,
+                code_frequency .* range ./ sampling_frequency .+ start_code_phase,
+                prn,
+            ) .* 10^(45 / 20) .+ randn(ComplexF64, length(range)) .* sqrt(4e6)
+        code = get_code.(
+            gpsl1,
+            code_frequency .* (-2:4001) ./ sampling_frequency .+ start_code_phase,
+            prn,
+        )
+        correlator = EarlyPromptLateCorrelator()
+        sample_shifts =
+            get_correlator_sample_shifts(correlator, sampling_frequency, code_frequency)
+        correlator = Tracking.downconvert_and_correlate_fused!(
+            correlator,
+            signal,
+            code,
+            sample_shifts,
+            0.0Hz,
+            sampling_frequency,
+            0.0,
+            start_sample,
+            num_samples,
+        )
+        cn0_estimator = update(cn0_estimator, get_prompt(correlator))
+    end
+    @test @inferred(get_current_index(cn0_estimator)) == 100
+    @test @inferred(Base.length(cn0_estimator)) == 100
+    cn0_estimate = @inferred estimate_cn0(cn0_estimator, 1ms)
+
+    @test cn0_estimate ≈ 45dBHz atol = 1.0dBHz
+end
+
+end
