+++ test/runtests.jl
@@ -22,7 +22,10 @@ include("galileo_e5a.jl")
 include("bit_buffer.jl")
 include("bit_integration_test.jl")
 include("correlator.jl")
-include("cn0_estimation.jl")
+include("cn0_estimators/cn0_estimator.jl")
+include("cn0_estimators/moments.jl")
+include("cn0_estimators/no_cn0.jl")
+include("cn0_estimators/nwpr.jl")
 include("cn0_estimator_comparison.jl")
 include("tracking_state.jl")
 include("track.jl")
